Initial commit and support 3.6 or greater
Migrates the API out of the yunwah/tennoui repository. The
warframe_api.py file is renamed to warpy.py and the API class has been
renamed to Worldstate. This rename came about as a result of what parts
of the API were implemented. Next steps may include the static data
processing part of the API as its own class. Additionally this commit
aims to make the library installable however there are no current test
files so it may fail. From testing with the interface after installing
this current state of warpy works as is. The following are changes
made to the API wrapper:

Exceptions - used for non valid platforms and languages.
Language headers - API now supports getting back results in different
languages, the default being en.