adds the "state" value to the returned open world json payloads
This option adds the "state" value which is meant to make the json payloads
easier to understand for open worlds. Currently vallisCycle does not have a
value that provides a string based value of its current state where as
cetusCycle and cambionCycle do. One of the other issues is that there is no
consistency between cetusCycle and cambionCycle for their string based
representation of the worlds open state. cetusCycle uses 'state' where as
cambionCycle uses 'active'. In their own respective in-game terms this isn't
necessarily wrong.